web/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic>=2.0
python-multipart
pyyaml
